Added package and a second menu
No action listener yet, but the menu works, at least
