--- MavenClock/src/main/java/ClockMain.java
@@ -1,17 +1,0 @@
-
-import java.awt.EventQueue;
-
-public class ClockMain {
-	public static void main(String[] args) {
-		
-		EventQueue.invokeLater(new Runnable() {
-			public void run() {
-				try {
-					ClockWindow frame = new ClockWindow();
-					} catch (Exception e) {
-					e.printStackTrace();
-				}
-			}
-		});
-	}
-}
